Rollback users.hooks.js, inserted code into authentication.js instead
